+++ app/src/main/java/com/nate/autofinance/domain/usecases/period/SelectPeriodUseCase.kt
@@ -3,7 +3,7 @@ import com.nate.autofinance.data.repository.PeriodRepository
 class SelectPeriodUseCase(
     private val periodRepo: PeriodRepository
 ) {
-    suspend operator fun invoke(periodId: Int, userId: Long) {
+    suspend operator fun invoke(periodId: Int, userId: Int) {
         val all = periodRepo.getPeriodsForUser(userId)
         all.forEach {
             val updated = it.copy(isSelected = (it.id == periodId))
